Fix maybe unitialised warning
When trying to compile using -O3 GCC detected that there is a variable
in use which might not be initialised. Here the solution is not so
obvious and plenty of other suspicious chunks of code lie around here,
e.g. passing a sint as u32.

However, using uninitialised values would directly result in undefined
behaviour, so its legal to choose any value one wants to. I chose 0,
which I suspect to be the most sane option.