Fix build for kernel 4.19
Two issues had to be fixed for kernel 4.19.

The first was a define of a formerly defined symbol. This is resolved
almost free of risk by guarding the definition by #ifndef.

The second issue was that a function signature did not match what was
expected. This was naively adapted, since the value in question was not
used anyways.